requirements.txt
faiss-cpu
langchain
openai
tiktoken
python-dotenv

